Move fields merge and unmerge logic to FieldRowViewModel for better testability
