+++ src/main/java/org/jabref/gui/mergeentries/newmergedialog/FieldRowView.java
@@ -1,22 +1,15 @@
 package org.jabref.gui.mergeentries.newmergedialog;
 
-import javax.swing.undo.AbstractUndoableEdit;
-import javax.swing.undo.CannotRedoException;
-import javax.swing.undo.CannotUndoException;
-import javax.swing.undo.CompoundEdit;
-
 import javafx.beans.property.ReadOnlyStringProperty;
 import javafx.scene.control.ToggleGroup;
 import javafx.scene.layout.GridPane;
 
-import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.gui.mergeentries.newmergedialog.cell.FieldNameCell;
 import org.jabref.gui.mergeentries.newmergedialog.cell.FieldValueCell;
 import org.jabref.gui.mergeentries.newmergedialog.cell.MergedFieldCell;
 import org.jabref.gui.mergeentries.newmergedialog.cell.sidebuttons.ToggleMergeUnmergeButton;
 import org.jabref.gui.mergeentries.newmergedialog.diffhighlighter.SplitDiffHighlighter;
 import org.jabref.gui.mergeentries.newmergedialog.diffhighlighter.UnifiedDiffHighlighter;
-import org.jabref.gui.mergeentries.newmergedialog.fieldsmerger.FieldMerger;
 import org.jabref.gui.mergeentries.newmergedialog.fieldsmerger.FieldMergerFactory;
 import org.jabref.gui.mergeentries.newmergedialog.toolbar.ThreeWayMergeToolbar;
 import org.jabref.model.entry.BibEntry;
@@ -43,10 +36,8 @@ public class FieldRowView {
 
     private final FieldRowViewModel viewModel;
 
-    private final CompoundEdit fieldsMergedEdit = new CompoundEdit();
-
     public FieldRowView(Field field, BibEntry leftEntry, BibEntry rightEntry, BibEntry mergedEntry, FieldMergerFactory fieldMergerFactory, int rowIndex) {
-        viewModel = new FieldRowViewModel(field, leftEntry, rightEntry, mergedEntry);
+        viewModel = new FieldRowViewModel(field, leftEntry, rightEntry, mergedEntry, fieldMergerFactory);
 
         fieldNameCell = new FieldNameCell(field.getDisplayName(), rowIndex);
         leftValueCell = new FieldValueCell(viewModel.getLeftFieldValue(), rowIndex);
@@ -61,9 +52,9 @@ public class FieldRowView {
             EasyBind.listen(toggleMergeUnmergeButton.fieldStateProperty(), ((observableValue, old, fieldState) -> {
                 LOGGER.debug("Field merge state is {} for field {}", fieldState, field);
                 if (fieldState == ToggleMergeUnmergeButton.FieldState.MERGED) {
-                    new MergeCommand(fieldMergerFactory.create(field)).execute();
+                    viewModel.mergeFields();
                 } else {
-                    new UnmergeCommand().execute();
+                    viewModel.unmergeFields();
                 }
             }));
         }
@@ -180,69 +171,5 @@ public class FieldRowView {
         int rightValueLength = getRightValueCell().getStyleClassedLabel().getLength();
         getRightValueCell().getStyleClassedLabel().clearStyle(0, rightValueLength);
         getRightValueCell().getStyleClassedLabel().replaceText(viewModel.getRightFieldValue());
-    }
-
-    public class MergeCommand extends SimpleCommand {
-        private final FieldMerger fieldMerger;
-
-        public MergeCommand(FieldMerger fieldMerger) {
-            this.fieldMerger = fieldMerger;
-
-            this.executable.bind(viewModel.hasEqualLeftAndRightBinding().not());
-        }
-
-        @Override
-        public void execute() {
-            assert !viewModel.getLeftFieldValue().equals(viewModel.getRightFieldValue());
-
-            String oldLeftFieldValue = viewModel.getLeftFieldValue();
-            String oldRightFieldValue = viewModel.getRightFieldValue();
-
-            String mergedFields = fieldMerger.merge(viewModel.getLeftFieldValue(), viewModel.getRightFieldValue());
-            viewModel.setLeftFieldValue(mergedFields);
-            viewModel.setRightFieldValue(mergedFields);
-
-            if (fieldsMergedEdit.canRedo()) {
-                fieldsMergedEdit.redo();
-            } else {
-                fieldsMergedEdit.addEdit(new MergeFieldsUndo(oldLeftFieldValue, oldRightFieldValue, mergedFields));
-                fieldsMergedEdit.end();
-            }
-        }
-    }
-
-    public class UnmergeCommand extends SimpleCommand {
-        @Override
-        public void execute() {
-            if (fieldsMergedEdit.canUndo()) {
-                fieldsMergedEdit.undo();
-            }
-        }
-    }
-
-    class MergeFieldsUndo extends AbstractUndoableEdit {
-        private final String oldLeft;
-        private final String oldRight;
-        private final String mergedFields;
-
-        MergeFieldsUndo(String oldLeft, String oldRight, String mergedFields) {
-            this.oldLeft = oldLeft;
-            this.oldRight = oldRight;
-            this.mergedFields = mergedFields;
-        }
-
-        @Override
-        public void undo() throws CannotUndoException {
-            super.undo();
-            viewModel.setLeftFieldValue(oldLeft);
-            viewModel.setRightFieldValue(oldRight);
-        }
-
-        @Override
-        public void redo() throws CannotRedoException {
-            super.redo();
-            viewModel.setLeftFieldValue(mergedFields);
-            viewModel.setRightFieldValue(mergedFields);
-        }
     }
 }
